Added confirmation modal

See merge request tekkom/web/queenbee!9

+++ src/app/pages/events/data-table-event/data-table-event.component.ts
@@ -56,7 +56,9 @@ export class DataTableEventComponent implements AfterViewInit {
   }
 
   onDelete(id: number): void {
-    this.dataSource.deleteItem(id);
-    this.dataSource.refresh();
+    if(confirm("Are you sure to delete the event with id: " + id)) {
+      this.dataSource.deleteItem(id);
+      this.dataSource.refresh();
+    }
   }
 }
