+++ src/app/pages/locations/data-table-location/data-table-location.component.ts
@@ -50,7 +50,9 @@ export class DataTableLocationComponent implements AfterViewInit {
   }
 
   onDelete(id: number): void {
-    this.dataSource.deleteItem(id);
-    this.dataSource.refresh();
+    if(confirm("Are you sure to delete the event with id: " + id)) {
+      this.dataSource.deleteItem(id);
+      this.dataSource.refresh();
+    }
   }
 }
